SDG-App/server/requirements.lock.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.lock.txt requirements.in
#
alembic==1.18.4
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==25.1.0
    # via passlib
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.3.1
    # via uvicorn
cryptography==44.0.1
    # via
    #   -r requirements.in
    #   python-jose
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.3.0
    # via
    #   -r requirements.in
    #   pydantic
fastapi==0.128.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.3.0
    # via pytest
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
packaging==26.0
    # via pytest
passlib[argon2]==1.7.4
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.23
    # via cffi
pydantic[email]==2.10.5
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via -r requirements.in
pymysql==1.1.1
    # via -r requirements.in
pytest==9.0.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-jose[cryptography]==3.4.0
    # via -r requirements.in
python-multipart==0.0.22
    # via -r requirements.in
pyyaml==6.0.3
    # via uvicorn
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sqlalchemy==2.0.45
    # via
    #   -r requirements.in
    #   alembic
starlette==0.50.0
    # via
    #   -r requirements.in
    #   fastapi
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic-settings
uvicorn[standard]==0.32.1
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
